fix test

+++ tests/test_align/test_template/test_vision.py
@@ -256,7 +256,7 @@ def test_molmoe():
 
 
 def test_doc_owl2():
-    pt_engine = PtEngine('iic/DocOwl2')
+    pt_engine = PtEngine('iic/DocOwl2', torch_dtype=torch.float16)
     images = [
         'https://modelscope.cn/models/iic/DocOwl2/resolve/master/examples/docowl2_page0.png',
         'https://modelscope.cn/models/iic/DocOwl2/resolve/master/examples/docowl2_page1.png',
